requirements.txt
# 核心依赖
openpyxl>=3.1.0
python-docx>=0.8.11

# Web API 依赖
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
slowapi>=0.1.9

# 可选：生产部署
gunicorn>=21.0.0
